feat: refactor signal handling and add completion percentage calculation for income models
