+++ income_tax_returns/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'income_tax_returns.apps.IncomeTaxReturnsConfig'
