+++ income_tax_returns/apps.py
@@ -2,5 +2,10 @@ from django.apps import AppConfig
 
 
 class IncomeTaxReturnsConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
     name = 'income_tax_returns'
+    default_auto_field = 'django.db.models.BigAutoField'
+
+    def ready(self):
+        import income_tax_returns.signals  # 👈 ensure this line exists
+
+
