+++ income_tax_returns/models.py
@@ -5,7 +5,7 @@ from djongo.models import ArrayField, EmbeddedField, JSONField
 from .helpers import *
 from usermanagement.models import ServiceRequest, Users
 from servicetasks.models import ServiceTask
-from django.db.models.signals import post_save
+from django.db.models.signals import post_save, pre_save
 from django.dispatch import receiver
 from docwallet.models import PrivateS3Storage
 
@@ -1338,286 +1338,3 @@ class ReviewFilingCertificate(models.Model):
     def __str__(self):
         return self.review_certificate_status or "No Review Status"
 
-
-def calculate_completion_percentage(instance, exclude_fields=None):
-    exclude_fields = exclude_fields or ['id', 'created_at', 'updated_at', 'service_request', 'service_task']
-    total_fields = 0
-    filled_fields = 0
-
-    for field in instance._meta.fields:
-        if field.name in exclude_fields:
-            continue
-        total_fields += 1
-        value = getattr(instance, field.name)
-        if value not in [None, '', []]:
-            filled_fields += 1
-
-    if total_fields == 0:
-        return 0
-    return round((filled_fields / total_fields) * 100)
-
-
-@receiver(post_save, sender=PersonalInformation)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
-
-
-@receiver(post_save, sender=TaxPaidDetails)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
-
-
-@receiver(post_save, sender=SalaryIncome)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
-
-
-@receiver(post_save, sender=OtherIncomeDetails)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
-
-
-@receiver(post_save, sender=NRIEmployeeSalaryDetails)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
-
-
-@receiver(post_save, sender=HousePropertyIncomeDetails)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
-
-
-@receiver(post_save, sender=CapitalGainsApplicableDetails)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
-
-
-@receiver(post_save, sender=CapitalGainsEquityMutualFund)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
-
-
-@receiver(post_save, sender=OtherCapitalGains)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
-
-
-@receiver(post_save, sender=DividendIncome)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
-
-
-@receiver(post_save, sender=GiftIncomeDetails)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
-
-
-@receiver(post_save, sender=BusinessProfessionalIncome)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
-
-
-@receiver(post_save, sender=InterestIncome)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
-
-
-@receiver(post_save, sender=FamilyPensionIncome)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
-
-
-@receiver(post_save, sender=ForeignIncome)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
-
-
-@receiver(post_save, sender=WinningIncome)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
-
-
-@receiver(post_save, sender=AgricultureIncome)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
-
-
-@receiver(post_save, sender=Deductions)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
-
-
-@receiver(post_save, sender=ReviewFilingCertificate)
-def sync_service_task_status(sender, instance, **kwargs):
-    task = instance.service_task
-
-    # Sync status
-    if task.status != instance.status:
-        task.status = instance.status
-
-    # Sync completion %
-    task.completion_percentage = calculate_completion_percentage(instance)
-
-    task.save()
